cube-client/tests/test_basic.py
import pytest
import sys

# Mock hardware for testing
sys.modules["RPi"] = __import__("fake_rpi.RPi", fromlist=[""])
sys.modules["RPi.GPIO"] = __import__("fake_rpi.RPi.GPIO", fromlist=[""])

def test_cube_basic():
    """Basic test for cube client"""
    assert True

def test_five_button_layout():
    """Test 5-button configuration"""
    import fake_rpi.RPi.GPIO as GPIO
    
    GPIO.setmode(GPIO.BCM)
    
    # Test 5-button layout
    buttons = {
        "A": 17,      # Top-left corner
        "B": 18,      # Top-right corner  
        "C": 22,      # Bottom-right corner
        "D": 23,      # Bottom-left corner
        "CENTER": 24  # Center generate button
    }
    
    for button_name, pin in buttons.items():
        GPIO.setup(pin, GPIO.IN, pull_up_down=GPIO.PUD_UP)
    
    assert len(buttons) == 5
    print(f"✅ 5-button layout configured: {list(buttons.keys())}")

@pytest.mark.asyncio
async def test_cube_async():
    """Test async cube functionality"""
    await asyncio.sleep(0.001)
    assert True
